feat(QOV-799): add kubeconfig and bastion setting in load-credentials command
